# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Campus Resource Optimization Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: resources
  - name: bookings
  - name: analytics
  - name: iot
  - name: ai
  - name: notifications
  - name: chatbot
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthStatus" }

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List all resources with optional filters
      parameters:
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: building
          schema: { type: string }
      responses:
        "200":
          description: Resources list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Resource" }

  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a single resource
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resource" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /resources/{id}/availability:
    get:
      operationId: getResourceAvailability
      tags: [resources]
      summary: Get hourly availability for a resource on a date
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: query
          name: date
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AvailabilitySlot" }

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (filterable)
      parameters:
        - in: query
          name: userId
          schema: { type: integer }
        - in: query
          name: resourceId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string }
        - in: query
          name: to
          schema: { type: string }
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (with conflict detection)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "409":
          description: Conflict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConflictResponse" }

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
    delete:
      operationId: cancelBooking
      tags: [bookings]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }

  /bookings/{id}/checkin:
    post:
      operationId: checkInBooking
      tags: [bookings]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }

  /bookings/calendar:
    get:
      operationId: getCalendar
      tags: [bookings]
      summary: All bookings for calendar view
      parameters:
        - in: query
          name: from
          schema: { type: string }
        - in: query
          name: to
          schema: { type: string }
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CalendarEvent" }

  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: KPI summary cards
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AnalyticsSummary" }

  /analytics/utilization:
    get:
      operationId: getUtilizationByType
      tags: [analytics]
      summary: 7-day utilization grouped by resource type
      responses:
        "200":
          description: Utilization data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UtilizationByType" }

  /analytics/peakhours:
    get:
      operationId: getPeakHours
      tags: [analytics]
      summary: Hourly usage heatmap (7 days x 14 hours)
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HeatmapCell" }

  /analytics/trends:
    get:
      operationId: getBookingTrends
      tags: [analytics]
      summary: 30-day booking trends
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrendPoint" }

  /analytics/underused:
    get:
      operationId: getUnderusedResources
      tags: [analytics]
      summary: Resources with low utilization
      responses:
        "200":
          description: Underused resources
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UnderusedResource" }

  /analytics/energy:
    get:
      operationId: getEnergyUsage
      tags: [analytics]
      summary: Energy usage per building
      responses:
        "200":
          description: Energy data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EnergyPoint" }

  /analytics/topresources:
    get:
      operationId: getTopResources
      tags: [analytics]
      summary: Top 5 most booked resources
      responses:
        "200":
          description: Top resources
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TopResource" }

  /analytics/role-breakdown:
    get:
      operationId: getRoleBreakdown
      tags: [analytics]
      summary: Bookings by role (faculty vs student)
      responses:
        "200":
          description: Role breakdown
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RoleBreakdown" }

  /iot/sensors:
    get:
      operationId: listSensors
      tags: [iot]
      summary: All IoT sensor readings
      responses:
        "200":
          description: Sensors
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SensorReading" }

  /iot/simulate:
    post:
      operationId: simulateIoT
      tags: [iot]
      summary: Trigger random occupancy update
      responses:
        "200":
          description: Updated readings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SensorReading" }

  /ai/predict:
    get:
      operationId: predictDemand
      tags: [ai]
      summary: Hourly demand prediction for a resource
      parameters:
        - in: query
          name: resourceId
          required: true
          schema: { type: integer }
        - in: query
          name: date
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Demand scores
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemandPrediction" }

  /ai/suggest:
    get:
      operationId: suggestSlots
      tags: [ai]
      summary: Suggest optimal booking slots
      parameters:
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: capacity
          schema: { type: integer }
        - in: query
          name: date
          schema: { type: string }
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SlotSuggestion" }

  /ai/insights:
    get:
      operationId: getInsights
      tags: [ai]
      summary: AI-generated insights for dashboard
      responses:
        "200":
          description: Insights
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AIInsight" }

  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - in: query
          name: userId
          schema: { type: integer }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }

  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Notification" }

  /notifications/read-all:
    post:
      operationId: markAllRead
      tags: [notifications]
      parameters:
        - in: query
          name: userId
          schema: { type: integer }
      responses:
        "200":
          description: All marked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CountResponse" }

  /chatbot:
    post:
      operationId: chat
      tags: [chatbot]
      summary: Send message to campus assistant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChatRequest" }
      responses:
        "200":
          description: Reply
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatResponse" }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]

    CountResponse:
      type: object
      properties:
        count: { type: integer }
      required: [count]

    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [token, user]

    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, faculty, student] }
        department: { type: string }
        avatar: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, name, email, role, department, createdAt]

    Resource:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        building: { type: string }
        floor: { type: integer }
        capacity: { type: integer }
        amenities:
          type: array
          items: { type: string }
        lat: { type: number }
        lng: { type: number }
        status: { type: string }
        currentOccupancy: { type: integer }
        energyUsage: { type: number }
        createdAt: { type: string, format: date-time }
      required: [id, name, type, building, floor, capacity, amenities, lat, lng, status, currentOccupancy, energyUsage, createdAt]

    AvailabilitySlot:
      type: object
      properties:
        hour: { type: integer }
        available: { type: boolean }
        bookingId: { type: integer }
        bookingTitle: { type: string }
      required: [hour, available]

    Booking:
      type: object
      properties:
        id: { type: integer }
        resourceId: { type: integer }
        resourceName: { type: string }
        userId: { type: integer }
        userName: { type: string }
        userRole: { type: string }
        title: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        date: { type: string }
        attendees: { type: integer }
        status: { type: string }
        priority: { type: integer }
        checkInTime: { type: string, format: date-time }
        qrCode: { type: string }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, resourceId, resourceName, userId, userName, userRole, title, startTime, endTime, date, attendees, status, priority, qrCode, createdAt]

    CreateBookingRequest:
      type: object
      properties:
        resourceId: { type: integer }
        userId: { type: integer }
        title: { type: string }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        attendees: { type: integer }
        notes: { type: string }
        override: { type: boolean }
      required: [resourceId, userId, title, startTime, endTime, attendees]

    ConflictResponse:
      type: object
      properties:
        error: { type: string }
        conflicts:
          type: array
          items: { $ref: "#/components/schemas/Booking" }
        canOverride: { type: boolean }
      required: [error, conflicts, canOverride]

    CalendarEvent:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        resourceId: { type: integer }
        resourceName: { type: string }
        userName: { type: string }
        userRole: { type: string }
        status: { type: string }
        priority: { type: integer }
      required: [id, title, start, end, resourceId, resourceName, userName, userRole, status, priority]

    AnalyticsSummary:
      type: object
      properties:
        totalResources: { type: integer }
        activeBookingsToday: { type: integer }
        utilizationPct: { type: number }
        energySavedKwh: { type: number }
        totalBookings: { type: integer }
        cancelledPct: { type: number }
        avgDurationMinutes: { type: number }
      required: [totalResources, activeBookingsToday, utilizationPct, energySavedKwh, totalBookings, cancelledPct, avgDurationMinutes]

    UtilizationByType:
      type: object
      properties:
        day: { type: string }
        classroom: { type: number }
        lab: { type: number }
        study_room: { type: number }
        parking: { type: number }
        equipment: { type: number }
      required: [day, classroom, lab, study_room, parking, equipment]

    HeatmapCell:
      type: object
      properties:
        day: { type: string }
        hour: { type: integer }
        utilization: { type: number }
      required: [day, hour, utilization]

    TrendPoint:
      type: object
      properties:
        date: { type: string }
        bookings: { type: integer }
        cancellations: { type: integer }
      required: [date, bookings, cancellations]

    UnderusedResource:
      type: object
      properties:
        resourceId: { type: integer }
        name: { type: string }
        type: { type: string }
        utilizationPct: { type: number }
        lastBooked: { type: string }
        suggestion: { type: string }
      required: [resourceId, name, type, utilizationPct, suggestion]

    EnergyPoint:
      type: object
      properties:
        building: { type: string }
        date: { type: string }
        kwh: { type: number }
        wastedKwh: { type: number }
      required: [building, date, kwh, wastedKwh]

    TopResource:
      type: object
      properties:
        resourceId: { type: integer }
        name: { type: string }
        type: { type: string }
        bookingCount: { type: integer }
      required: [resourceId, name, type, bookingCount]

    RoleBreakdown:
      type: object
      properties:
        role: { type: string }
        count: { type: integer }
      required: [role, count]

    SensorReading:
      type: object
      properties:
        resourceId: { type: integer }
        resourceName: { type: string }
        capacity: { type: integer }
        currentOccupancy: { type: integer }
        occupancyPct: { type: number }
        status: { type: string }
        updatedAt: { type: string, format: date-time }
      required: [resourceId, resourceName, capacity, currentOccupancy, occupancyPct, status, updatedAt]

    DemandPrediction:
      type: object
      properties:
        resourceId: { type: integer }
        date: { type: string }
        hourly:
          type: array
          items:
            type: object
            properties:
              hour: { type: integer }
              score: { type: number }
            required: [hour, score]
        peakHours:
          type: array
          items: { type: integer }
      required: [resourceId, date, hourly, peakHours]

    SlotSuggestion:
      type: object
      properties:
        resourceId: { type: integer }
        resourceName: { type: string }
        startHour: { type: integer }
        endHour: { type: integer }
        date: { type: string }
        matchScore: { type: number }
        reason: { type: string }
      required: [resourceId, resourceName, startHour, endHour, date, matchScore, reason]

    AIInsight:
      type: object
      properties:
        id: { type: string }
        icon: { type: string }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        action: { type: string }
        actionLabel: { type: string }
      required: [id, type, title, message]

    Notification:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        message: { type: string }
        type: { type: string }
        read: { type: boolean }
        relatedBookingId: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, userId, message, type, read, createdAt]

    ChatRequest:
      type: object
      properties:
        message: { type: string }
        userId: { type: integer }
      required: [message]

    ChatResponse:
      type: object
      properties:
        reply: { type: string }
        intent: { type: string }
        suggestions:
          type: array
          items: { type: string }
        actions:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              value: { type: string }
            required: [label, value]
      required: [reply, intent]