# lib/api-zod/src/generated/types/availabilitySlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Resource Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AvailabilitySlot {
  hour: number;
  available: boolean;
  bookingId?: number;
  bookingTitle?: string;
}