# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Resource Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  resourceId: number;
  userId: number;
  title: string;
  startTime: Date;
  endTime: Date;
  attendees: number;
  notes?: string;
  override?: boolean;
}