# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Resource Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Resource {
  id: number;
  name: string;
  type: string;
  building: string;
  floor: number;
  capacity: number;
  amenities: string[];
  lat: number;
  lng: number;
  status: string;
  currentOccupancy: number;
  energyUsage: number;
  createdAt: Date;
}