# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Resource Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  date: string;
  bookings: number;
  cancellations: number;
}