# lib/api-zod/src/generated/types/underusedResource.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Campus Resource Optimization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UnderusedResource {
  resourceId: number;
  name: string;
  type: string;
  utilizationPct: number;
  lastBooked?: string;
  suggestion: string;
}